+++ module/config.js
@@ -1995,6 +1995,8 @@ export const PF1 = {
       "mod",
     ],
   },
+
+  keepItemLinksOnCopy: ["classAssociations"],
 };
 
 /**