+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Copying an attack with linked ammunition to another actor made it behave strangely (#801)
 - Fix error on rolling concentration with no ability associated with the spellbook (#825)
 - `@abilities.<abl>.base` and `@abilities.<abl>.baseMod` weren't being updated properly (#824)
+- Class Associations weren't being copied over anymore
 
 ### Compendium